Move ExtensionInstaller to Composer\EventListener\PackageEventListener
